+++ src/tests/integration/test_postgreSQL.py
@@ -6,7 +6,6 @@ def test_meta_cmd_display(ip_with_postgreSQL):
     assert ("public", "taxi", "table", "ploomber_app") in out.result
 
 
-# Known issue, addressing in https://github.com/ploomber/jupysql/issues/90
 def test_auto_commit_mode_on(ip_with_postgreSQL, capsys):
     ip_with_postgreSQL.run_cell("%config SqlMagic.autocommit=True")
     ip_with_postgreSQL.run_cell("%sql CREATE DATABASE new_db")