+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # CHANGELOG
 
-## 0.5.7dev
+## 0.6.1dev
 * [Feature] Adds `%sqlcmd profile` (#66)
 
+## 0.6.0 (2023-02-27)
+
+* [API Change] Drops support for old versions of IPython (removed imports from `IPython.utils.traitlets`)
 * [Feature] Adds `%%config SqlMagic.autopolars = True` ([#138](https://github.com/ploomber/jupysql/issues/138))
 
 ## 0.5.6 (2023-02-16)